new: Fail hoc
