+++ src/components/card-details/card-details.e2e.test.js
@@ -4,7 +4,6 @@ import Adapter from 'enzyme-adapter-react-16';
 import CardDetails from './card-details';
 import {offer} from '../../test-data/offer-test-data';
 import {ComponentType} from '../../utils/const';
-import {Link, MemoryRouter} from 'react-router-dom';
 configure({adapter: new Adapter()});
 
 describe(`CardDetails tests`, () => {
