+++ src/components/locations-item/locations-item.e2e.test.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import {configure, shallow} from 'enzyme';
 import {LocationsItem} from './locations-item';
-import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
 configure({adapter: new Adapter()});
