+++ src/components/map-section/map-section.test.js
@@ -3,14 +3,14 @@ import renderer from 'react-test-renderer';
 import MapSection from './map-section';
 import {offers} from '../../test-data/offers-test-data';
 
-  it(`should render MapSection correctly`, () => {
-    const tree = renderer
+it(`should render MapSection correctly`, () => {
+  const tree = renderer
       .create(
-              <MapSection
-                offersToShowOnMap={offers}
-                activeCity={`Paris`}
-                activeOffer={0}
-              />,
+          <MapSection
+            offersToShowOnMap={offers}
+            activeCity={`Paris`}
+            activeOffer={0}
+          />,
           {
             createNodeMock: () => {
               return document.createElement(`div`);
@@ -18,5 +18,5 @@ import {offers} from '../../test-data/offers-test-data';
           }
       ).toJSON();
 
-    expect(tree).toMatchSnapshot();
-  });
+  expect(tree).toMatchSnapshot();
+});
