+++ src/components/star-buttons/star-buttons.test.js
@@ -5,7 +5,7 @@ import {StarButtons} from './star-buttons';
 const noop = () => {};
 test(`Render StarButtons`, () => {
   const tree = renderer
-    .create(<StarButtons 
+    .create(<StarButtons
       onChange={noop}
       onStarClick={noop}
     />)
